Move `assert_image_mostly_same` helper to a shared module and update imports in tests
